Do not delete NOVA kernel sources if we already have them
This simplifies working offline. Only nul-local branch is
automatically reset; all other local branches are kept untouched.

+++ kernel/contrib/checkout-nova.sh
@@ -4,10 +4,18 @@ TARGET=$1
 URL=$2
 REVISION=$3
 
-rm -rf "$TARGET" || exit 1
-git clone "$URL" $TARGET || exit 1
+if test -d "$TARGET" && ( cd "$TARGET" && git rev-parse --is-inside-work-tree >/dev/null ) && \
+   test -z "$(cd "$TARGET" && git rev-parse --show-prefix)"; then
+    if ! ( cd "$TARGET" && git show --oneline $REVISION >/dev/null 2>&1 ); then
+	( cd "$TARGET" && git fetch )
+    fi
+else
+    rm -rf "$TARGET" || exit 1
+    git clone "$URL" $TARGET || exit 1
+    ( cd $TARGET && git branch nul-local ) || exit 1
+fi
 
-( cd $TARGET && git checkout -b nul-local "$REVISION" ) || exit 1
+( cd $TARGET && git checkout nul-local && git reset --hard "$REVISION" ) || exit 1
 
 shift 3
